third_party/WebKit/LayoutTests/media/video-controls-dont-show-on-focus-when-disabled.html

<!DOCTYPE html>
<title>Test that video controls don't appear when the media element is focused and the "controls" attribute is not set.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="media-controls.js"></script>
<!-- "tabindex" makes the video focusable despite it not having controls -->
<video tabindex="0"></video>
<script>
async_test(function(t) {
    var video = document.querySelector("video");

    video.onplaying = t.step_func(function() {
        runAfterHideMediaControlsTimerFired(t.step_func_done(function() {
            var controls = mediaControlsButton(video, "panel");
            video.focus();
            assert_equals(getComputedStyle(controls).opacity, "0");
        }), video);
    });

    video.src = "content/test.ogv";
    video.play();
});
</script>
